refactor: move exceptions to dedicated 'types' package
